Initial commit to use "pdb" when available for check-syntax
This is a squash of a dozen initial commits. As part of that, I had
started to make changes to the status quo racket-xp-mode and
check-syntax.rkt, before realizing I would use a new, dedicated
racket-pdb-mode. Then I changed racket-xp.el and check-syntax.rkt back
-- somewhat but not exactly. This complicate rebasing this branch onto
master. So revert racket-xp.el and check-syntax.rkt to what's on the
main branch.

As part of this squash, I'm losing some early history even wrt to just
racket-pdb.el and pdb.rkt -- but I think that's fine: It was very
exploratory and going to get squashed someday anyway.